[#189] test: 로직 분리에 따른 관련 테스트 패키지 이동
